requirements.txt
aiohttp==3.7.3
asn1crypto==1.4.0
async-timeout==3.0.1
attrs==18.2.0
certifi==2018.8.24
cffi==1.14.4
chardet==3.0.4
cryptography==2.3.1
decorator==4.0.11
Django==2.2.13
future==0.16.0
idna==2.7
idna-ssl==1.1.0
image==1.5.25
imageio==2.1.2
InstagramAPI==1.0.2
multidict==5.0.2
numpy==1.19.4
peewee==3.7.0
Pillow==8.0.1
pkg-resources==0.0.0
pycparser==2.19
python-decouple==3.1
python-telegram-bot==11.1.0
pytz==2018.5
requests==2.11.1
requests-toolbelt==0.7.0
runp==0.0.2
six==1.11.0
sqlparse==0.4.1
telegram==0.0.1
telepot==12.7
tqdm==4.11.2
typing-extensions==3.7.4.3
urllib3==1.24.2
yarl==1.6.3
